Update Bot Balance Alarms
This commit allows for the creation of a new class for both the disputer bot and liquidator bot when the index.js file is run. This commit also adds all of the recommended amendments provided by Suraj's feedback.
